backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
psycopg2-binary
python-dotenv
alembic
pydantic
python-multipart
